Make livewire extend LayerGroupPointer, add tap behavior
